Revert "Missed a float16"
This reverts commit 282379ee0b396630b3c8a9c0c6cab88bf38dc638.

+++ rl/common/utils.py
@@ -60,7 +60,7 @@ numpy_to_torch_dtype_dict = {
 
 
 def construct_blank_tensors(size, transition_shapes):
-    return [torch.zeros(size, *shp, dtype=torch.float16) for shp in transition_shapes]
+    return [torch.zeros(size, *shp) for shp in transition_shapes]
 
 
 def tensor(x):
